chore(deps): update: server side layout
